+++ Vixen.System/Data/Value/PositionValue.cs
@@ -4,7 +4,7 @@ namespace Vixen.Data.Value
 {
 	public struct PositionValue : IIntentDataType
 	{
-		public PositionValue(double percentage)
+		public PositionValue(float percentage)
 		{
 			if (percentage < 0 || percentage > 1) throw new ArgumentOutOfRangeException("percentage");
 
@@ -14,6 +14,6 @@ namespace Vixen.Data.Value
 		/// <summary>
 		/// Percentage value between 0 and 1.
 		/// </summary>
-		public double Position;
+		public float Position;
 	}
 }
